backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
langchain==0.2.6
langchain-groq==0.1.6
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.3.0
pydantic==2.7.4
sqlalchemy==2.0.30
asyncpg==0.29.0
boto3==1.34.131
pytest==8.2.2
httpx==0.27.0
alembic==1.13.1
minio==7.2.7
psycopg2-binary==2.9.9